requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
blobfile==3.0.0
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
chroma-hnswlib==0.7.6
chromadb==1.0.10
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
configparser==7.2.0
crcmod==1.7
cryptography==45.0.3
curl_cffi==0.11.1
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
durationpy==0.9
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.9
fastjsonschema==2.21.1
filelock==3.18.0
flatbuffers==25.2.10
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2025.3.2
google-auth==2.39.0
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kubernetes==32.0.1
langchain==0.3.25
langchain-chroma==0.2.4
langchain-community==0.3.24
langchain-core==0.3.61
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
langgraph==0.4.7
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.1
langgraph-sdk==0.1.70
langsmith==0.3.42
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.4.4
multitasking==0.0.11
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.2.4
oauthlib==3.2.2
onnxruntime==1.21.0
openai==1.82.0
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-asgi==0.53b1
opentelemetry-instrumentation-fastapi==0.53b1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opentelemetry-util-http==0.53b1
orjson==3.10.18
ormsgpack==1.10.0
oss2==2.19.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
peewee==3.18.1
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.8
posthog==3.25.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.4
psycopg2==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pycryptodomex==3.23.0
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pymongo==4.13.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.45.3
sympy==1.13.3
tenacity==9.1.2
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
tomli==2.2.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.1
watchfiles==1.0.5
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
xxhash==3.5.0
yarg==0.1.9
yarl==1.20.0
yfinance==0.2.61
zipp==3.21.0
zstandard==0.23.0
